refactor: Restructure
