+++ crates/pretty-exec-lib/src/lib.rs
@@ -1,10 +1,10 @@
+pub mod app;
 pub mod logger;
-pub mod main;
 pub mod pretty_exec;
 pub mod sub;
 
+pub use app::*;
 pub use logger::*;
-pub use main::*;
 pub use pretty_exec::*;
 
 pub use structopt_utilities::{self, clap, structopt};
